Rename { -> FunctionalTest}BucketProvider
and move `split` out of `BucketProvider`.
