+++ .teamcity/Gradle_Check/model/FunctionalTestBucketProvider.kt
@@ -42,13 +42,13 @@ val CROSS_VERSION_BUCKETS = listOf(
 
 typealias BuildProjectToSubprojectTestClassTimes = Map<String, Map<String, List<TestClassTime>>>
 
-interface GradleBuildBucketProvider {
+interface FunctionalTestBucketProvider {
     fun createFunctionalTestsFor(stage: Stage, testCoverage: TestCoverage): List<FunctionalTest>
 
     fun createDeferredFunctionalTestsFor(stage: Stage): List<FunctionalTest>
 }
 
-class StatisticBasedGradleBuildBucketProvider(private val model: CIBuildModel, testTimeDataJson: File) : GradleBuildBucketProvider {
+class StatisticBasedFunctionalTestBucketProvider(private val model: CIBuildModel, testTimeDataJson: File) : FunctionalTestBucketProvider {
     private val buckets: Map<TestCoverage, List<BuildTypeBucket>> = buildBuckets(testTimeDataJson, model)
 
     override fun createFunctionalTestsFor(stage: Stage, testCoverage: TestCoverage): List<FunctionalTest> {
@@ -167,46 +167,6 @@ class StatisticBasedGradleBuildBucketProvider(private val model: CIBuildModel, t
     }
 }
 
-/**
- * Split a list of elements into nearly even sublist. If an element is too large, largeElementSplitFunction will be used to split the large element into several smaller pieces;
- * if some elements are too small, they will be aggregated by smallElementAggregateFunction.
- *
- * @param list the list to split, must be ordered by size desc
- * @param toIntFunction the function used to map the element to its "size"
- * @param largeElementSplitFunction the function used to further split the large element into smaller pieces
- * @param smallElementAggregateFunction the function used to aggregate tiny elements into a large bucket
- * @param expectedBucketNumber the return value's size should be expectedBucketNumber
- */
-fun <T, R> split(list: LinkedList<T>, toIntFunction: (T) -> Int, largeElementSplitFunction: (T, Int) -> List<R>, smallElementAggregateFunction: (List<T>) -> R, expectedBucketNumber: Int, maxNumberInBucket: Int, noElementSplitFunction: (Int) -> List<R> = { throw IllegalArgumentException("More buckets than things to split") }): List<R> {
-    if (expectedBucketNumber == 1) {
-        return listOf(smallElementAggregateFunction(list))
-    }
-    if (list.isEmpty()) {
-        return noElementSplitFunction(expectedBucketNumber)
-    }
-
-    val expectedBucketSize = list.sumBy(toIntFunction) / expectedBucketNumber
-
-    val largestElement = list.removeFirst()!!
-
-    val largestElementSize = toIntFunction(largestElement)
-
-    return if (largestElementSize >= expectedBucketSize) {
-        val bucketsOfFirstElement = largeElementSplitFunction(largestElement, if (largestElementSize % expectedBucketSize == 0) largestElementSize / expectedBucketSize else largestElementSize / expectedBucketSize + 1)
-        val bucketsOfRestElements = split(list, toIntFunction, largeElementSplitFunction, smallElementAggregateFunction, expectedBucketNumber - bucketsOfFirstElement.size, maxNumberInBucket, noElementSplitFunction)
-        bucketsOfFirstElement + bucketsOfRestElements
-    } else {
-        val buckets = arrayListOf(largestElement)
-        var restCapacity = expectedBucketSize - toIntFunction(largestElement)
-        while (restCapacity > 0 && list.isNotEmpty() && buckets.size < maxNumberInBucket) {
-            val smallestElement = list.removeLast()
-            buckets.add(smallestElement)
-            restCapacity -= toIntFunction(smallestElement)
-        }
-        listOf(smallElementAggregateFunction(buckets)) + split(list, toIntFunction, largeElementSplitFunction, smallElementAggregateFunction, expectedBucketNumber - 1, maxNumberInBucket, noElementSplitFunction)
-    }
-}
-
 class GradleVersionRangeCrossVersionTestBucket(private val startInclusive: String, private val endExclusive: String) : BuildTypeBucket {
     override fun createFunctionalTestsFor(model: CIBuildModel, stage: Stage, testCoverage: TestCoverage, bucketIndex: Int) =
         FunctionalTest(model,
@@ -220,7 +180,7 @@ class GradleVersionRangeCrossVersionTestBucket(private val startInclusive: Strin
         )
 }
 
-class LargeSubprojectSplitBucket(val subproject: GradleSubproject, val number: Int, val include: Boolean, val classes: List<TestClassTime>) : BuildTypeBucket by subproject {
+class LargeSubprojectSplitBucket(val subproject: GradleSubproject, number: Int, val include: Boolean, val classes: List<TestClassTime>) : BuildTypeBucket by subproject {
     val name = "${subproject.name}_$number"
     val totalTime = classes.sumBy { it.buildTimeMs }
