+++ .teamcity/Gradle_Check/settings.kts
@@ -1,7 +1,7 @@
 package Gradle_Check
 
 import Gradle_Check.model.JsonBasedGradleSubprojectProvider
-import Gradle_Check.model.StatisticBasedGradleBuildBucketProvider
+import Gradle_Check.model.StatisticBasedFunctionalTestBucketProvider
 import jetbrains.buildServer.configs.kotlin.v2019_2.project
 import jetbrains.buildServer.configs.kotlin.v2019_2.version
 import model.CIBuildModel
@@ -30,5 +30,5 @@ calling the subProjects() method in this project.
 
 version = "2020.1"
 val model = CIBuildModel(buildScanTags = listOf("Check"), subprojects = JsonBasedGradleSubprojectProvider(File("./subprojects.json")))
-val gradleBuildBucketProvider = StatisticBasedGradleBuildBucketProvider(model, File("./test-class-data.json"))
+val gradleBuildBucketProvider = StatisticBasedFunctionalTestBucketProvider(model, File("./test-class-data.json"))
 project(RootProject(model, gradleBuildBucketProvider))
